Implement array initializers.
This is a little bit of a hack.
Real C# will use a RuntimeFieldHandle object, which cannot be properly serialized.
First, we extract the raw bytes of the array initializer in the inline token call.
Finally, we replace invocations of RuntimeHelpers.InitializeArray with a helper function that uses Buffer.BlockCopy.